improved versioning for developer version
